+++ src/Oedering/Ordering.Application/Features/Orders/Commands/CheckoutOrder/CheckoutOrderCommand.cs
@@ -1,9 +1,4 @@
 ﻿using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Ordering.Application.Features.Orders.Commands.CheckoutOrder
 {
@@ -12,9 +7,9 @@ namespace Ordering.Application.Features.Orders.Commands.CheckoutOrder
         public string UserName { get; set; }
         public string CoinName { get; set; }
         public int CoinId { get; set; }
-        public decimal PriceCoin { get; set; }
-        public decimal Amount { get; set; }
-        public decimal TotalPrice { get; set; }
+        public double PriceCoin { get; set; }
+        public double Amount { get; set; }
+        public double TotalPrice { get; set; }
         public DateTime? DateTime { get; set; }
