+++ src/Oedering/Ordering.Application/Features/Orders/Commands/UpdateOrder/UpdadeOrderCommand.cs
@@ -13,9 +13,9 @@ namespace Ordering.Application.Features.Orders.Commands.UpdateOrder
         public string UserName { get; set; }
         public string CoinName { get; set; }
         public int CoinId { get; set; }
-        public decimal PriceCoin { get; set; }
-        public decimal Amount { get; set; }
-        public decimal TotalPrice { get; set; }
+        public double PriceCoin { get; set; }
+        public double Amount { get; set; }
+        public double TotalPrice { get; set; }
 
         // BillingAddress
         public string FirstName { get; set; }
